Extract auth functions to their own class
