+++ api/DHSC.FingertipsNext.Modules.DataManagement/Controllers/V1/DataManagementBatchController.cs
@@ -1,4 +1,3 @@
-using System.Security.Claims;
 using DHSC.FingertipsNext.Modules.Common.Auth;
 using DHSC.FingertipsNext.Modules.Common.Schemas;
 using DHSC.FingertipsNext.Modules.DataManagement.Schemas;
@@ -31,32 +30,6 @@ public class DataManagementBatchController : ControllerBase
         _dataManagementService = dataManagementService;
     }
 
-    private static List<Guid> GetRoles(ClaimsPrincipal user)
-    {
-        var userRoleIds = user.FindAll(ClaimTypes.Role).Select(c => c.Value).ToArray();
-        var userRoles = new List<Guid>();
-
-        foreach (var id in userRoleIds)
-            if (Guid.TryParse(id, out var roleGuid))
-            {
-                userRoles.Add(roleGuid);
-            }
-
-        return userRoles;
-    }
-
-    private static string? GetUserId(ClaimsPrincipal user)
-    {
-        return user.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-    }
-
-    private async Task<int[]> GetIndicatorIdsFromRoles(IEnumerable<Guid> roles)
-    {
-        var userIndicatorPermissions = await _indicatorPermissionsLookupService.GetIndicatorsForRoles(roles);
-
-        return userIndicatorPermissions.ToArray();
-    }
-
     /// <summary>
     ///     Get details of all health data upload batches that are for indicators that you have permissions to modify.
     /// </summary>
@@ -68,8 +41,8 @@ public class DataManagementBatchController : ControllerBase
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> ListBatches()
     {
-        var userRoles = GetRoles(User);
-        var indicatorIds = await GetIndicatorIdsFromRoles(userRoles);
+        var userRoles = AuthUtilities.GetRoles(User);
+        var indicatorIds = await AuthUtilities.GetIndicatorIdsFromRoles(_indicatorPermissionsLookupService, userRoles);
 
         if (indicatorIds.Length == 0 && !User.IsInRole(_adminRole))
         {
@@ -98,8 +71,8 @@ public class DataManagementBatchController : ControllerBase
             });
         }
 
-        var userRoles = GetRoles(User);
-        var indicatorIds = await GetIndicatorIdsFromRoles(userRoles);
+        var userRoles = AuthUtilities.GetRoles(User);
+        var indicatorIds = await AuthUtilities.GetIndicatorIdsFromRoles(_indicatorPermissionsLookupService, userRoles);
         if (indicatorIds.Length == 0 && !User.IsInRole(_adminRole))
         {
             return new ForbidResult();
@@ -111,7 +84,7 @@ public class DataManagementBatchController : ControllerBase
             indicatorIds = [];
         }
 
-        var userId = GetUserId(User);
+        var userId = AuthUtilities.GetUserId(User);
         if (userId == null)
         {
             return new ForbidResult();
